frontend/requirements.txt
streamlit
pymupdf
openai
azure-cosmos
azure-storage-blob